+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes
 1.3.1 (unreleased)
 ------------------
 
+- Getting language by acquisition, indeed some Dexterity content has no language or empty language.
+  [bsuttor]
+
 - Fix empty context.REQUEST.get('form.widgets.taxonomy'). It add it into POST form.
   [bsuttor]