+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changes
 =======
 
 
-1.3.1 (unreleased)
+1.4.0 (2016-11-03)
 ------------------
 
 - Getting language by acquisition, indeed some Dexterity content has no language or empty language.